# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Ordering & Payment System API
  description: |
    Production-grade e-commerce backend API built with Express, TypeScript,
    Prisma ORM, Redis-backed category caching, and modular services.

    Authentication is handled through JWT bearer tokens.
    Payment providers include Stripe and bKash.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/hasanemam1717/E-commerce-Ordering-Payment-System
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.production.example.com/api
    description: Production

# ─────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (future — not yet enforced).

  # ─── Reusable Schemas ──────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: Resource not found
        context:
          type: object
          description: Additional error context (e.g. productIds, available)
          example:
            productId: 'uuid-here'
            requested: 999
            available: 10
      examples:
        notFound:
          value:
            status: error
            message: Resource not found
        validationError:
          value:
            status: error
            message: Unprocessable entity
        rateLimited:
          value:
            status: error
            message: Too many requests, please try again later.

    # ─── Enums ───────────────────────────────────────────────
...
    OrderStatus:
      type: string
      enum: [PENDING, PAID, CANCELED]
...
    CreateOrderRequest:
      type: object
      required: [userId, items]
      properties:
        userId:
          type: string
          format: uuid
          example: 48843698-2ea6-49eb-afe5-045e9fd5e7a7
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
                format: uuid
                example: 8cf0df89-4a4e-417e-9046-249222eac9cf
              quantity:
                type: integer
                minimum: 1
                example: 2
    OrderItemResponse:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        productName:
          type: string
          example: MacBook Pro 14"
        sku:
          type: string
          example: LAP-001
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          format: decimal
          example: 1999.99
        subtotal:
          type: number
          format: decimal
          example: 3999.98
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        total:
          type: number
          format: decimal
          example: 4699.96
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt:
          type: string
          format: date-time

    # ─── Payment / Webhook ──────────────────────────────────
...
  responses:
...
    400BadRequest:
      description: Bad request — invalid payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
      examples:
        unsupportedProvider:
          value:
            status: error
            message: "Unsupported payment provider: 'PAYPAL'"
            context:
              provider: PAYPAL
...
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
      examples:
        categoryNotFound:
          value:
            status: error
            message: "Category with id 'bad-id' not found"
        orderNotFound:
          value:
            status: error
            message: "Payment 'bad-id' not found"
            context:
              paymentId: bad-id
    409Conflict:
      description: Resource conflict (e.g. insufficient stock)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
      examples:
        insufficientStock:
          value:
            status: error
            message: "Insufficient stock for product 'p-macbook': requested 999, available 10"
            context:
              productId: p-macbook
              requested: 999
              available: 10
    422UnprocessableEntity:
      description: Validation or business logic error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
      examples:
        inactiveProduct:
          value:
            status: error
            message: "Product 'p-xyz' is not active (current status: INACTIVE)"
            context:
              productId: p-xyz
              status: INACTIVE
        invalidState:
          value:
            status: error
            message: "Order 'order-1' is in status 'CANCELED' and cannot be paid"
            context:
              orderId: order-1
              status: CANCELED
...
    500InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
      examples:
        internalError:
          value:
            status: error
            message: Internal server error

# ═════════════════════════════════════════════════════════════
#  Paths
# ═════════════════════════════════════════════════════════════
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Create a new order with validated line items.
      description: >
        Creates an order with PENDING status. Validates that all products
        exist, are ACTIVE, and have sufficient stock. Uses an interactive
        Prisma transaction with FOR UPDATE row locking to prevent race
        conditions and overselling. All monetary values use Prisma.Decimal
        for floating-point-safe arithmetic.
      operationId: createOrder
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              userId: '48843698-2ea6-49eb-afe5-045e9fd5e7a7'
              items:
                - productId: '8cf0df89-4a4e-417e-9046-249222eac9cf'
                  quantity: 1
                - productId: 'ccf78956-c179-4f2c-b712-06faf876d9cb'
                  quantity: 2
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
              example:
                id: 49d402a7-90a0-48e7-86ad-d07d8698faed
                userId: '48843698-2ea6-49eb-afe5-045e9fd5e7a7'
                status: PENDING
                total: 2699.97
                items:
                  - productId: 8cf0df89-4a4e-417e-9046-249222eac9cf
                    productName: MacBook Pro 14"
                    sku: LAP-001
                    quantity: 1
                    unitPrice: 1999.99
                    subtotal: 1999.99
                  - productId: ccf78956-c179-4f2c-b712-06faf876d9cb
                    productName: Sony WH-1000XM5
                    sku: HDN-001
                    quantity: 2
                    unitPrice: 349.99
                    subtotal: 699.98
                createdAt: '2026-07-22T09:40:06.228Z'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          $ref: '#/components/responses/404NotFound'
        '409':
          $ref: '#/components/responses/409Conflict'
        '422':
          $ref: '#/components/responses/422UnprocessableEntity'
        '500':
          $ref: '#/components/responses/500InternalServerError'

  # ─── Payments ──────────────────────────────────────────────
...